added custom column search
