--- src/library/components/table/column-search/column-search-input/types.ts
@@ -1,9 +1,0 @@
-import { ColumnSearch } from "../../types"
-
-export interface GenericColumnSearchInputProps {
-  value: string | undefined
-  accessor: string | number
-  columnSearch: ColumnSearch
-  onColumnSearchInput: (newValue: string, column: string | number) => void
-  onSubmit: () => void
-}
